Установили зависимость '--cov'
