+++ pyproject.toml
@@ -44,6 +44,7 @@ zstandard = "==0.23.0"
 [tool.poetry.group.dev.dependencies]
 requests = "^2.32.3"
 pytest = "^8.3.5"
+pytest-cov = "^6.1.1"
 
 
 [tool.poetry.group.lint.dependencies]
